python/requirements.txt
# PDF-Compare Core Dependencies
# Only includes libraries needed for PDF comparison (no Flask/GUI)
numpy>=2.4.1
opencv-python-headless>=4.11.0.86
pdf2image>=1.17.0
pdfplumber>=0.11.9
pillow>=12.1.0
